refactor: eliminate typings module; relocate types to domain-appropriate modules
- Move ParsedSortingQuery from typings/ to dependencies/sorting_query.py
- Replace typings.Token TypedDict with existing schemas.Token Pydantic model
- Update security.create_token() to return Pydantic Token instances
- Fix token access in tests (dict access -> attribute access)
- Remove unused PaginatedResponse type
- Delete entire typings/ directory
- Update all import statements to reference new type locations

This refactor improves code organization by co-locating types with their
domain logic, eliminates redundant type definitions, and follows FastAPI
conventions for response models.
