--- src/kwik/typings/sorting_query.py
@@ -1,5 +1,0 @@
-"""Type definitions for query sorting."""
-
-from __future__ import annotations
-
-ParsedSortingQuery = list[tuple[str, str]]
